web/requirements.txt
asgiref==3.3.0
attrs==20.2.0
autobahn==20.7.1
Automat==20.2.0
bcrypt==3.2.0
certifi==2020.6.20
cffi==1.14.3
channels==2.4.0
click==7.1.2
constantly==15.1.0
cryptography==3.1.1
daphne==2.5.0
Django==3.1
django-captcha==1.0.0
django-ranged-response==0.2.0
django-recaptcha==2.0.6
django-simple-email-confirmation==0.70
hyperlink==20.0.1
idna==2.10
incremental==17.5.0
opentelemetry-api==0.11b0
opentelemetry-ext-asgi==0.11b0
opentelemetry-ext-dbapi==0.11b0
opentelemetry-ext-django==0.11b1
opentelemetry-ext-jaeger==0.11b1
opentelemetry-ext-psycopg2==0.11b1
opentelemetry-ext-wsgi==0.11b0
opentelemetry-instrumentation==0.11b0
opentelemetry-sdk==0.11b0
pipdeptree==1.0.0
psycopg2-binary==2.8.6
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
PyHamcrest==2.0.2
pyOpenSSL==19.1.0
python-dotenv==0.14.0
pytz==2020.1
redis==3.5.3
rq==1.5.2
sentry-sdk==0.19.0
service-identity==18.1.0
six==1.15.0
sqlparse==0.4.1
thrift==0.13.0
Twisted==20.3.0
txaio==20.4.1
urllib3==1.25.10
wrapt==1.12.1
zope.interface==5.1.2
